refactoring project_structure
