+++ Sumdays/app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:theme="@style/Theme.First" />
         <!-- Sum 화면 -->
         <activity
-            android:name=".SumActivity"
+            android:name=".DailySumActivity"
             android:exported="false" />
         <!-- Calendar 화면 (실제 첫 화면) -->
         <activity
