+++ Sumdays/app/src/main/res/layout/activity_sum.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/sum_layout"
-    tools:context=".SumActivity">
+    tools:context=".DailySumActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/header_layout"
@@ -82,7 +82,7 @@
 
     <include
         android:id="@+id/bottom_buttons_layout"
-        layout="@layout/navigation_bar_calendar"
+        layout="@layout/include_nav_calendar"
         android:layout_width="0dp"
         android:layout_height="150dp"
         app:layout_constraintTop_toBottomOf="@id/sum_button_divider"
