+++ Sumdays/app/src/main/res/layout/include_nav_daily.xml
@@ -64,11 +64,11 @@
     </LinearLayout>
     <ImageButton
         android:id="@+id/btnSum"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
+        android:layout_width="90dp"
+        android:layout_height="90dp"
         android:src="@drawable/ic_fox"
         android:background="@android:color/transparent"
-        android:scaleType="fitXY"
+        android:scaleType="fitCenter"
         android:layout_gravity="center_horizontal|bottom"
         android:translationY="-85dp"
         app:tint="@null"/> </FrameLayout>
